form added to page, added css
