+++ src/components/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import '../styles/App.css';
 
-import applicantForm from './Form';
+import ApplicantForm from './ApplicantForm';
 import applicantResponse from './Response'
 
 class App extends Component {
@@ -9,8 +9,13 @@ class App extends Component {
     return (
       <div className="App">
         <header className="App-header">
-
+          <h1> FoodLogiQ Coding Challenge! </h1>
+          <p> Submit answers to the questions below.</p>
+          <ApplicantForm />
         </header>
+        <article>
+
+        </article>
       </div>
     );
   }
