+++ src/styles/form.css
@@ -1,0 +1,16 @@
+form {
+  display: flex;
+  flex-direction: column;
+  max-width: 85%;
+  justify-content: center;
+  margin: 0 auto;
+}
+
+button {
+  margin: 2rem auto;
+}
+
+.label {
+  padding-top: 2rem;
+  text-align: left;
+}
